linker: try remaining locations on library load failure
This commit modifies the behavior of the linker when trying to load
a library. Until now, the linker stopped on the first file it could
open in the path search order (starting with LD_LIBRARY_PATH); if
loading the file failed for any reason, then linking would fail. Now,
the linker tries to load the file immediately after opening it; if
either opening or loading the file fails, then it moves on to the next
location.  If all attempts fail, they are all reported in the same
order, otherwise nothing is reported. Some error messages have been
modified in order to always show the full path to the file that could
not be loaded.

This new behavior matches that of ld-linux.so. It is particularly
useful in that it allows to support search paths with mixed bitness
(libraries with the wrong bitness are merely ignored). More
specifically, it enables 64b executables to execute 32b files (or the
other way around) with LD_LIBRARY_PATH containing both 32b and 64b
paths, which is impossible with the old behavior.

AOSPA Ticket: NOUGAT-35

Change-Id: Ieedd08e4f5bfb3b17bcd07af2d091ed03210ba81
Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>