document that fields before cached_pid break it
Adding a field above this results in either that field or the tid field
(depending on if it's above that too) being set to zero by vfork.
Ideally that assembly would be replaced with a C++ wrapper but in the
meantime this adds a comment to save time for someone else.

Signed-off-by: anupritaisno1 <www.anuprita804@gmail.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>