libm: Add optimized expf2
This implementation is based on code from ARM optimized routines [1].
The error for expf, for nearest rouding mode, is estimated on 0.502 ulp
with worst case in non-nearest up to 1 ulp.

Double precision arithmetics is used with internal calculation and it
is expected to be faster and easier to get better precision on most
targets.  The const data is kept in a separate translation unit that
allows fast code for literal loads on most targets and also allows
sharing data across expf, exp2f and powf.

The performance shows about 1.3x better throughput with same latency
for aarch64 (A72).

Test: ran 32-bit and 64-bit x86 tests on host
Change-Id: I1a086b86940fa1902ff04c9b28e2d9525445fff4
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>