Use --Bsymbolic-functions on libm
This option removes mostly of the PLT calls for internal symbol, they
are routed to internal entrypoint instead.  It should be safe for libm,
since symbol interposing is not supported.

For arm64 the remaining PLTs are __cxa_finalize, ldexp, __stack_chk_fail,
isnanf, memset, and isxdigit.  All of them are expected (clang (r339432)
does not have optimization for isnan/isinf inline expansion and these
symbols are defined on libc.so).

Test: Builds, ran unit tests on arm and x86.

Change-Id: I1eb1a9f2b30f21788d9ad8639b4a0a15f6d0566a
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>