Allow larger page sizes on host Bionic on arm64.
We cannot rely on the page size being 4096 on non-Android
Linux kernels on arm64 as the maximum page size is 65536 on that
architecture. Therefore, update Bionic and Soong to allow page sizes
up to 65536 on host Bionic platforms on arm64. The page size remains
4096 on Android.

Most of the code in Bionic needs to know what the actual page size
is at runtime, not just the maximum page size. Therefore, introduce
an inline page_size() function that will either return the runtime
page size (on non-Android arm64) or a constant 4096 (elsewhere). This
should ensure that there are no changes to the generated code on
unaffected platforms.

Bug: 230790254
Change-Id: Ic0ed98b67f7c6b845804b90a4e16649f2fc94028
Signed-off-by: minaripenguin37 <alexfinhart@gmail.com>
Signed-off-by: DennySPb <dennyspb@gmail.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>